+++ lib/kitchen/directions/move_solutions_to_answer_key/answer_key_inner_container.rb
@@ -1,32 +1,40 @@
 # frozen_string_literal: true
 
 module Kitchen::Directions::AnswerKeyInnerContainer
-  def self.v1(chapter:, metadata_source:, append_to:, solutions_plural: true, in_appendix: false)
+  def self.v1(chapter:, metadata_source:, append_to:, options: {
+    solutions_plural: true,
+    in_appendix: false,
+    cases: false
+  })
+    options.reverse_merge!(
+      solutions_plural: true,
+      in_appendix: false,
+      cases: false
+    )
     V1.new.bake(
       chapter: chapter,
       metadata_source: metadata_source,
       append_to: append_to,
-      solutions_plural: solutions_plural,
-      in_appendix: in_appendix
+      options: options
     )
   end
 
   class V1
     renderable
 
-    def bake(chapter:, metadata_source:, append_to:, solutions_plural:, in_appendix:)
-      @solutions_or_solution = solutions_plural ? 'solutions' : 'solution'
-      @uuid_key = if in_appendix
+    def bake(chapter:, metadata_source:, append_to:, options:)
+      @solutions_or_solution = options[:solutions_plural] ? 'solutions' : 'solution'
+      @uuid_key = if options[:in_appendix]
                     "appendix#{@solutions_or_solution}#{chapter.count_in(:book)}"
                   else
                     @uuid_key = "#{@solutions_or_solution}#{chapter.count_in(:book)}"
                   end
       @metadata = metadata_source.children_to_keep.copy
       @composite_element = 'composite-page'
-      @title = if in_appendix
-                 "#{I18n.t(:appendix)} #{[*('A'..'Z')][chapter.count_in(:book) - 1]}"
+      @title = if options[:in_appendix]
+                 "#{I18n.t("appendix#{'.nominative' if options[:cases]}")} #{[*('A'..'Z')][chapter.count_in(:book) - 1]}"
                else
-                 "#{I18n.t(:chapter)} #{chapter.count_in(:book)}"
+                 "#{I18n.t("chapter#{'.nominative' if options[:cases]}")} #{chapter.count_in(:book)}"
                end
       @main_title_tag = 'h2'