+++ spec/kitchen_spec/directions/bake_appendix_spec.rb
@@ -3,6 +3,11 @@
 require 'spec_helper'
 
 RSpec.describe Kitchen::Directions::BakeAppendix do
+  before do
+    stub_locales({
+      'appendix': 'Appendix'
+    })
+  end
 
   let(:page) do
     page_element(
@@ -72,7 +77,7 @@ RSpec.describe Kitchen::Directions::BakeAppendix do
     expect(page_appendix_with_section_learning_objectives).to match_snapshot_auto
   end
 
-  context 'when book has added has adde target labels for appendices' do
+  context 'when book has added has added target labels for appendices and does not use gramatical cases' do
     it 'stores link text' do
       pantry = page.pantry(name: :link_text)
       expect(pantry).to receive(:store).with('Appendix 3 Appendix Title', { label: 'apId' })
@@ -84,7 +89,41 @@ RSpec.describe Kitchen::Directions::BakeAppendix do
     it 'stores link text' do
       pantry = page.pantry(name: :link_text)
       expect(pantry).not_to receive(:store).with('Appendix 3 Appendix Title', { label: 'apId' })
-      described_class.v1(page: page, number: '3', block_target_label: true)
+      described_class.v1(page: page, number: '3', options: { block_target_label: true })
+    end
+  end
+
+  context 'when book uses grammatical cases' do
+    it 'uses nominative case in title' do
+      with_locale(:pl) do
+        stub_locales({
+          'appendix': {
+            'nominative': 'Dodatek',
+            'genitive': 'Dodatku'
+          }
+        })
+
+        described_class.v1(page: page, number: 3, options: { cases: true })
+        expect(page).to match_snapshot_auto
+      end
+    end
+
+    it 'stores link text' do
+      with_locale(:pl) do
+        stub_locales({
+          'appendix': {
+            'nominative': 'Dodatek',
+            'genitive': 'Dodatku'
+          }
+        })
+
+        pantry = page.pantry(name: :nominative_link_text)
+        expect(pantry).to receive(:store).with('Dodatek 3 Appendix Title', { label: 'apId' })
+
+        pantry = page.pantry(name: :genitive_link_text)
+        expect(pantry).to receive(:store).with('Dodatku 3 Appendix Title', { label: 'apId' })
+        described_class.v1(page: page, number: '3', options: { cases: true })
+      end
     end
   end
 end